Moving Navigation Controller
We were misusing the navigation controller/navigation stack. I've placed the Login and the EULA view controllers before the navigation stack as we don't want the user to have a back button to get to these screens. As such, we've also had to change the segues that were being used.